PMSX003: Add support for specifying the update interval and spinning down
The PMSX003 series includes a moving fan, and device lifetime is likely to
be limited by (a) how long the fan is running, and (b) how much dust the
device ingests. If continuous sensor readings aren't required, we can spin
the fan down and increase the expected useful lifetime of the sensor.
However, there's an expected delay between turning the fan on and getting
reasonable data back, so we only want to spin the fan down if the update
interval is greater than that. This involves maintaining some additional
state in the loop() method, but the core functionality remains consistent.